backend/pyproject.toml
[project]
name = "chainrisk-backend"
version = "0.1.0"
description = "ChainRisk - a concise, general-purpose collective intelligence engine for forecasting"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "ChainRisk Team" }
]

dependencies = [
    # Core web stack
    "flask>=3.0.0",
    "flask-cors>=6.0.0",

    # LLM client (OpenAI-compatible API)
    "openai>=1.0.0",

    # Zep Cloud
    "zep-cloud==3.13.0",

    # OASIS social simulation
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",

    # Documents and text
    "PyMuPDF>=1.24.0",
    # Encoding detection for non-UTF-8 text files
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "redis>=5.0.0",
    "psycopg[binary]>=3.2.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "plotly>=5.24.0",
    "networkx>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
